requirements.txt
channels==4.0.0
daphne==4.0.0
Django==4.2.3
fontawesomefree==6.4.0
openai==0.27.1
Pillow==10.0.0
python-dotenv==1.0.0
